pyproject.toml
[project]
name = "mission-ag3ents"
version = "0.1.0"
description = "AI Devs competition — task solutions and shared utilities"
requires-python = ">=3.12"
dependencies = [
    "openai",
    "httpx",
    "pydantic",
    "pydantic-settings",
    "python-dotenv",
    "colorama",
    "aiofiles>=25.1.0",
    "haversine>=2.9.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.41.0",
    "mcp[cli]>=1.26.0",
    "fastmcp>=3.1.1",
    "websockets>=16.0",
    "pillow>=12.2.0",
]

[tool.setuptools.packages.find]
include = ["common*", "task_*", "dashboard*"]

[project.optional-dependencies]
dev = [
    "ruff",
    "ipython",
]
